+++ lib/sidekiq/processor.rb
@@ -5,6 +5,7 @@ require 'sidekiq/middleware/server/active_record'
 require 'sidekiq/middleware/server/exception_handler'
 require 'sidekiq/middleware/server/unique_jobs'
 require 'sidekiq/middleware/server/failure_jobs'
+require 'sidekiq/middleware/server/logging'
 
 module Sidekiq
   class Processor
@@ -18,6 +19,7 @@ module Sidekiq
     def self.default_middleware
       Middleware::Chain.new do |m|
         m.add Middleware::Server::ExceptionHandler
+        m.add Middleware::Server::Logging
         m.add Middleware::Server::UniqueJobs
         m.add Middleware::Server::ActiveRecord
       end