+++ myapp/config/initializers/sidekiq.rb
@@ -1,6 +1,6 @@
 Sidekiq.configure_client do |config|
-  config.redis = Sidekiq::RedisConnection.create(:namespace => 'resque', :size => 2)
+  config.redis = { :namespace => 'resque', :size => 2 }
 end
 Sidekiq.configure_server do |config|
-  config.redis = Sidekiq::RedisConnection.create(:namespace => 'resque', :size => 25)
+  config.redis = { :namespace => 'resque', :size => 25 }
 end