+++ test/test_stats.rb
@@ -5,7 +5,8 @@ require 'sidekiq/processor'
 class TestStats < MiniTest::Unit::TestCase
   describe 'with redis' do
     before do
-      Sidekiq.redis = @redis = Sidekiq::RedisConnection.create(:url => 'redis://localhost/sidekiq_test')
+      Sidekiq.redis = { :url => 'redis://localhost/sidekiq_test' }
+      @redis = Sidekiq.redis
       @redis.flushdb
     end