+++ Changes.md
@@ -1,6 +1,10 @@
 HEAD
 -----------
 
+- Remove :namespace and :server CLI options (mperham)
+- Add ExceptionNotifier support (masterkain)
+- Add capistrano support (mperham)
+- Workers now log upon start and finish (mperham)
 - Messages for terminated workers are now automatically requeued (mperham)
 - Add support for Exceptional error reporting (bensie)